transcirber-app/requirements.txt
openai-whisper==20240930
pydub==0.25.1
ffmpeg-python==0.2.0
tqdm==4.67.1
transformers==4.52.2
sentencepiece==0.2.0